Use CGFloat instead of float for MessagesView
- Use CGFloat and do mix with float

+++ XBMC Remote/MessagesView.h
@@ -15,7 +15,7 @@
     CGFloat slideHeight;
 }
 
-- (id)initWithFrame:(CGRect)frame deltaY:(float)deltaY deltaX:(float)deltaX;
+- (id)initWithFrame:(CGRect)frame deltaY:(CGFloat)deltaY deltaX:(CGFloat)deltaX;
 - (void)showMessage:(NSString *)message timeout:(float)timeout color:(UIColor *)color;
 
 @property (nonatomic, retain) UILabel *viewMessage;
